+++ gcc/testsuite/gcc.target/i386/pr97971.c
@@ -5,7 +5,7 @@
 int
 foo (void)
 {
-  register _Complex long a asm ("rax");
+  register _Complex long long a asm ("rax");
   register int b asm ("rdx");
   asm ("# %0 %1" : "=&r" (a), "=r" (b));	/* { dg-error "inconsistent operand constraints in an 'asm'" } */
   return a;
